Reorganize methods in RGBColor class
This commit rearranges some methods in the RGBColor class according to their function. The toHex method now returns a HexColor object instead of a string. The make and setRGB methods which were previously near the top are now placed after some getter methods. These modifications are made for the ease of locating methods and enhancing code readability.