Add toString method and refactor HexColor class
Added a toString method to the HexColor class for easily obtaining the hex string representation of the color. The class has also been refactored, with methods rearranged for a more logical ordering. Methods for getting and setting the hex string representation, as well as converting to an HSL color, have been added.